requirements.txt
python-telegram-bot>=20.0
pymongo>=4.0
requests>=2.25
dnspython>=2.0